Broke indexer code up to match other tools
