+++ reditools/index/__init__.py
@@ -1,3 +1,4 @@
 from .main import main
+from .rtindexer import RTIndexer
 
-__all__ = ('main',)
+__all__ = ('main', 'RTIndexer')
